Underline the lines with errors, add hover message
